Version 0.5.3

+++ Modules/Base/Systems/Peer.h
@@ -52,7 +52,7 @@ public:
 	RPCConfigurationParameter() {}
 	virtual ~RPCConfigurationParameter() {}
 
-	uint32_t databaseID = 0;
+	uint64_t databaseID = 0;
 	std::shared_ptr<RPC::Parameter> rpcParameter;
 	std::vector<uint8_t> data;
 	std::vector<uint8_t> partialData;
